requirements.txt
fastapi
uvicorn
pandas
simple-salesforce
openpyxl
python-dotenv